requirements.txt
requests
python-dotenv
google-generativeai
openai
gradio
pydantic
markdown2
xhtml2pdf
python-docx
